+++ src/Form/Element/CKEditor.php
@@ -12,4 +12,23 @@ class CKEditor extends Wysiwyg
     {
         parent::__construct($path, $label, 'ckeditor');
     }
+
+
+    /**
+     * Add Params Editor for Upload files
+     */
+    public function initialize()
+    {
+        parent::initialize();
+
+        $params = collect($this->getParameters());
+
+        if(!$params->has('uploadUrl')){
+            $this->parameters['uploadUrl'] = route('admin.ckeditor.upload');
+        }
+
+        if(!$params->has('filebrowserUploadUrl')){
+            $this->parameters['filebrowserUploadUrl'] = route('admin.ckeditor.upload');
+        }
+    }
 }